requirements.txt
python-fasthtml
sqlalchemy
psycopg2-binary
python-dotenv
uvicorn
python-multipart
